login

+++ src/models/product.js
@@ -14,7 +14,7 @@ const Product = (sequelize, Sequelize) => {
       allowNull: false
     },
     price: {
-      type: Sequelize.FLOAT(10, 2),
+      type: Sequelize.FLOAT,
       allowNull: false
     }
   })
